tests: Add PositiveMultiView tests
